move update_indices_after_given_line_number_if_necessary to csv_xls
